add rest query parameters to module execution remove params from modules that doesn’t use it
