+++ x/ownership/client/rest/query.go
@@ -15,11 +15,6 @@ func registerQueryRoutes(cliCtx context.CLIContext, r *mux.Router) {
 		"/ownership/list",
 		queryListHandlerFn(cliCtx),
 	).Methods(http.MethodGet)
-
-	r.HandleFunc(
-		"/ownership/parameters",
-		queryParamsHandlerFn(cliCtx),
-	).Methods(http.MethodGet)
 }
 
 func queryListHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
@@ -30,26 +25,6 @@ func queryListHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 		}
 
 		route := fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QueryList)
-		res, height, err := cliCtx.QueryWithData(route, nil)
-		if err != nil {
-			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-			return
-		}
-
-		cliCtx = cliCtx.WithHeight(height)
-		rest.PostProcessResponse(w, cliCtx, res)
-	}
-}
-
-func queryParamsHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
-		if !ok {
-			return
-		}
-
-		route := fmt.Sprintf("custom/%s/parameters", types.QuerierRoute)
-
 		res, height, err := cliCtx.QueryWithData(route, nil)
 		if err != nil {
 			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
